Changed line 155 and formatted the UI better
I changed "Current" to "New Order" to better reflect that a new purchase is being made, also the New order needs to still to have a method to finalize the order.

I think  my use of in.nextline() at 223 is wrong. which causes an error. might need a new input?